+++ forms/content-dashboard.php
@@ -12,6 +12,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 		include( COREACTIVITY_PATH . 'forms/content-dashboard-overall.php' );
 		include( COREACTIVITY_PATH . 'forms/content-dashboard-database.php' );
 		include( COREACTIVITY_PATH . 'forms/content-dashboard-statistics.php' );
+		include( COREACTIVITY_PATH . 'forms/content-dashboard-location.php' );
 
 		?>
     </div>